# backend/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CPIM-MES API",
    "description": "CPIM準拠の生産管理システムのREST API仕様。認証は JWT Bearer Token。\nログインで取得したトークンを Authorize ボタンから設定すれば、保護エンドポイントを試行できます。",
    "version": "1.0.0",
    "contact": {
      "name": "CPIM-MES"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "本サーバ"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VALIDATION"
          },
          "message": {
            "type": "string",
            "example": "validation failed"
          },
          "details": {}
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "admin123"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOi..."
          },
          "username": {
            "type": "string",
            "example": "admin"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "planner",
              "operator",
              "viewer"
            ]
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "ログイン",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "認証失敗",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "自分の情報",
        "responses": {
          "200": {
            "description": "成功"
          },
          "401": {
            "description": "未認証"
          }
        }
      }
    },
    "/items": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "品目一覧",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Item"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "品目登録",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "作成",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "400": {
            "description": "バリデーションエラー",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/items/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "品目取得",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "description": "存在しません"
          }
        }
      },
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "品目更新",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "品目削除",
        "responses": {
          "204": {
            "description": "削除"
          }
        }
      }
    },
...
    "/items/import": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "CSVインポート",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/mrp/run": {
      "post": {
        "tags": [
          "MRP"
        ],
        "summary": "MRP実行",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MRPRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MRPResult"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/crp/run": {
      "post": {
        "tags": [
          "CRP"
        ],
        "summary": "CRP実行",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CRPRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CapacityLoadRow"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/forecast/run": {
      "post": {
        "tags": [
          "Forecast"
        ],
        "summary": "需要予測実行 (SMA/EXPO)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/abc-analysis": {
      "get": {
        "tags": [
          "ABC"
        ],
        "summary": "ABC分析（年間使用金額ベース）",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ABCRow"
                  }
                }
              }
            }
          }
        },
        "description": "直近12か月の物理ISSUE数量 × 標準原価でABC分類します。ADJUST/RECEIPT/予約は使用量から除外します。",
        "parameters": [
          {
            "name": "asOf",
            "in": "query",
            "required": false,
            "description": "分析基準日。省略時はアプリの業務日。",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ]
      }
    },
    "/cycle-counts": {
      "get": {
        "tags": [
          "CycleCount"
        ],
        "summary": "サイクルカウント一覧",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "COUNTED",
                "RECONCILED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/cycle-counts/{id}/record": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "tags": [
          "CycleCount"
        ],
        "summary": "実測値を記録 (差異があれば在庫調整も自動起票)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "countedQty": {
                    "type": "number"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "成功"
          }
        }
      }
    },
    "/lots": {
      "get": {
        "tags": [
          "Lots"
        ],
        "summary": "ロット一覧 (残数付き)",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Lot"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Lots"
        ],
        "summary": "ロット登録",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Lot"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "作成"
          }
        }
      }
    },
...
    "/audit-log": {
      "get": {
        "tags": [
          "Audit"
        ],
        "summary": "監査ログ照会",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditEntry"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/forecast/runs": {
      "get": {
        "tags": [
          "Forecast"
        ],
        "summary": "List saved forecast versions",
        "parameters": [
          {
            "in": "query",
            "name": "itemId",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Forecast versions"
          }
        }
      }
    },
    "/forecast/runs/{id}": {
      "get": {
        "tags": [
          "Forecast"
        ],
        "summary": "Get forecast version and values",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Forecast run detail"
          }
        }
      }
    },
    "/forecast/runs/{id}/activate": {
      "post": {
        "tags": [
          "Forecast"
        ],
        "summary": "Activate a DRAFT forecast version",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Activated"
          }
        }
      }
    },
    "/forecast/runs/{id}/consumption": {
      "get": {
        "tags": [
          "Forecast"
        ],
        "summary": "Calculate customer-order forecast consumption",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Consumption buckets"
          }
        }
      }
    },
    "/forecast/runs/{id}/apply-to-mps": {
      "post": {
        "tags": [
          "Forecast"
        ],
        "summary": "Publish consumed demand from ACTIVE version to MPS",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated MPS count"
          }
        }
      }
    },
...
  }
}